fix: mock server tests would always try to connect to emulator
The mock server tests would always try to connect to the emulator if the
SPANNER_EMULATOR_HOST had been set. This would cause the tests to fail,
as the client would either send all requests to the emulator (if the emulator
was running), or to return an UNAVAILABLE exception when the emulator was not
running.