fix: mock server tests would always try to connect to emulator (#30)
* fix: mock server tests would always try to connect to emulator

The mock server tests would always try to connect to the emulator if the
SPANNER_EMULATOR_HOST had been set. This would cause the tests to fail,
as the client would either send all requests to the emulator (if the emulator
was running), or to return an UNAVAILABLE exception when the emulator was not
running.

* test: also run unit tests with emulator enabled

GitHub Actions should also execute the unit tests with the emulator enabled, to
ensure that the setup of the mock server tests are not disabled by the enabling
of the emulator.